backend/requirements.txt
# Core Flask Framework
Flask==2.3.2
Werkzeug==2.3.4

# Flask Extensions
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.19
alembic==1.11.1

# Database Drivers
pyodbc==5.1.0              # For SQL Server (optional)
psycopg2-binary==2.9.9     # For PostgreSQL (production recommended)
PyMySQL==1.1.0             # For MySQL (alternative)

# Forms and Validation
WTForms==3.0.1
email-validator==2.0.0

# Security
bcrypt==4.0.1
cryptography==41.0.3

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and API
requests==2.31.0
urllib3==2.0.4

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Date and Time
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.1.0

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Production Servers
gunicorn==21.2.0           # For Linux/Mac deployment (Render, Railway, Heroku)
waitress==3.0.0            # For Windows deployment

# Additional Utilities
click==8.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Email Support
secure-smtplib==0.1.1
